==> 2023_2_逐1小时风向风速和小时内极大风速0-48小时短期预报算法赛/0_data_processing.ipynb <==
import getpass
KAGGLE = False if getpass.getuser() == "Kaihua" else True

import os, copy, zipfile, re
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from tqdm.notebook import tqdm
from sklearn.metrics import roc_auc_score, f1_score, recall_score, precision_score

import sys
if KAGGLE:
    sys.path.append("/kaggle/input/2023zhuxiaoshifengxiang")
    DATA_PATH = '/kaggle/input/2023zhuxiaoshifengxiang'
else:
    sys.path.append("../code集合/models")
    sys.path.append("../code集合/feature_extraction")
    DATA_PATH = 'data'


SEED = 2023
MODEL_SAVE_PATH = 'model'
RESULT_SAVE_PATH = 'results'
if not os.path.exists(MODEL_SAVE_PATH):
    os.makedirs(MODEL_SAVE_PATH)
if not os.path.exists(RESULT_SAVE_PATH):
    os.makedirs(RESULT_SAVE_PATH)
cat_feats = []

with zipfile.ZipFile(os.path.join(DATA_PATH, "track3_train.zip"), 'r') as zfile:
    file_paths = zfile.namelist()
    wind_sample_list = list(filter(lambda x: re.match("track_3t/wind/.*csv", x)!=None, file_paths))
    feat_sample_list = list(filter(lambda x: re.match("track_3t/features/.*csv", x)!=None, file_paths))
    wind_sample_list.sort()
    feat_sample_list.sort()
    print('wind csv len', len(wind_sample_list))
    print('feat csv len', len(feat_sample_list))
    wind_data = []
    for item_wind_name in tqdm(wind_sample_list, desc='wind'):
        with zfile.open(item_wind_name, "r") as item:
            item_data = pd.read_csv(item)
            wind_data.append(item_data)

    feat_data = []
    for item_feat_name in tqdm(feat_sample_list, desc='feat'):
        with zfile.open(item_feat_name, "r") as item:
            item_data = pd.read_csv(item)
            feat_data.append(item_data)

wind_data = pd.concat(wind_data, axis=0)
feat_data = pd.concat(feat_data, axis=0)
data = pd.merge(wind_data, feat_data, on=['ID'], how='outer')
data.to_parquet(os.path.join(DATA_PATH, 'track3_train.parquet'), index=False)

with zipfile.ZipFile(os.path.join(DATA_PATH, "track3_a.zip"), 'r') as zfile:
    file_paths = zfile.namelist()
    wind_sample_list = list(filter(lambda x: re.match("track_3a/wind/.*csv", x)!=None, file_paths))
    feat_sample_list = list(filter(lambda x: re.match("track_3a/features/.*csv", x)!=None, file_paths))
    wind_sample_list.sort()
    feat_sample_list.sort()
    print('wind csv len', len(wind_sample_list))
    print('feat csv len', len(feat_sample_list))
    wind_data = []
    for item_wind_name in tqdm(wind_sample_list, desc='wind'):
        with zfile.open(item_wind_name, "r") as item:
            item_data = pd.read_csv(item)
            wind_data.append(item_data)

    feat_data = []
    for item_feat_name in tqdm(feat_sample_list, desc='feat'):
        with zfile.open(item_feat_name, "r") as item:
            item_data = pd.read_csv(item)
            feat_data.append(item_data)

# wind_data = pd.concat(wind_data, axis=0)
feat_data = pd.concat(feat_data, axis=0)
# data = pd.merge(wind_data, feat_data, on=['ID'], how='outer')
data = feat_data
data.to_parquet(os.path.join(DATA_PATH, 'track3_a.parquet'), index=False)
